+++ slimer37.github.io/src/Root.tsx
@@ -1,7 +1,7 @@
 import Header from './components/Header.tsx'
 import Footer from './components/Footer.tsx'
 
-import Home from './Home.tsx'
+import Home from './routes/Home.tsx'
 
 function Root() {
   return (
